added images

+++ scripts/load_files.sh
@@ -1,8 +1,23 @@
-pip3 install gdown
-gdown https://drive.google.com/uc?id=1TDOCNh2FoeaKMbR9Wc-1_MsOv-54a_IV
 if [ ! $(which unzip) ]
 then
     sudo apt-get install unzip
 fi
-unzip dataset.zip -d .
-rm -rf dataset.zip
+
+if [ ! $(which gdown) ]
+then
+    pip3 install gdown
+fi
+
+if [ ! -d dataset ]
+then
+    gdown https://drive.google.com/uc?id=1TDOCNh2FoeaKMbR9Wc-1_MsOv-54a_IV
+    unzip dataset.zip -d .
+    rm -rf dataset.zip
+fi
+
+if [ ! -d step2/seg_from_SAM/images ]
+then 
+    gdown https://drive.google.com/uc?id=1iXYl_YlGeaA1Z12QBRgqDxgc4ZAGtsJ9
+    unzip segmentations.zip -d .
+    rm -rf segmentations.zip
+fi
